Move to spectator to collect metrics (#706)
* Move to the spectator to collect metrics. Default binding is to NoopRegistry.
Refer to https://github.com/Netflix/spectator for more details.

+++ priam/src/test/java/com/netflix/priam/backup/FakeBackupFileSystem.java
@@ -143,7 +143,7 @@ public class FakeBackupFileSystem implements IBackupFileSystem {
     }
 
     @Override
-    public int getAWSSlowDownExceptionCounter() {
+    public long getAWSSlowDownExceptionCounter() {
         return 0;
     }
